+++ packages/ocean-react-formik/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.28.0](https://github.com/ocean-ds/ocean-web/compare/v1.27.3...v1.28.0) (2022-02-23)
+
+**Note:** Version bump only for package @useblu/ocean-react-formik
+
 ## [1.27.3](https://github.com/ocean-ds/ocean-web/compare/v1.27.2...v1.27.3) (2022-02-15)
 
 **Note:** Version bump only for package @useblu/ocean-react-formik